requirements.txt
moviepy==1.0.3
gTTS==2.5.3
numpy==2.1.3
Pillow==10.4.0
imageio-ffmpeg==0.5.1
